tests/samples/requirements.txt
# Intentionally pinned to known-vulnerable versions for scanner verification.
# DO NOT install these in a real environment.

# DEP-PY-DJANGO-001: Django < 3.2.25 / 4.2.11 / 5.0.3
Django==3.2.1

# DEP-PY-FLASK-001: Flask < 2.3.0 (Werkzeug RegexConverter ReDOS)
Flask==2.2.0

# DEP-PY-FASTAPI-001: FastAPI < 0.109.1
fastapi==0.100.0

# DEP-PY-REQUESTS-001: requests < 2.31.0
requests==2.28.0

# DEP-PY-URLLIB3-001: urllib3 < 1.26.17 / 2.0.4
urllib3==1.26.0

# DEP-PY-AIOHTTP-001: aiohttp < 3.9.2
aiohttp==3.8.0

# DEP-PY-JINJA2-001: Jinja2 < 3.1.3
Jinja2==3.1.0

# DEP-PY-PYYAML-001: PyYAML < 6.0
PyYAML==5.4.1

# DEP-PY-CRYPTOGRAPHY-001: cryptography < 41.0.6
cryptography==41.0.0

# DEP-PY-PARAMIKO-001: paramiko < 2.10.1
paramiko==2.8.0

# DEP-PY-PILLOW-001: Pillow < 10.2.0
Pillow==9.5.0

# DEP-PY-SQLALCHEMY-001: SQLAlchemy < 1.4.0
SQLAlchemy==1.3.23

# DEP-PY-CELERY-001: celery < 4.4.0
celery==4.3.0

# DEP-PY-LANGCHAIN-001: langchain < 0.0.312
langchain==0.0.300

# DEP-PY-TRANSFORMERS-001: transformers < 4.36.0
transformers==4.30.0

# DEP-PY-TORCH-001: torch < 2.0.0
torch==1.13.1

# DEP-PY-GRADIO-001: gradio < 3.41.0
gradio==3.40.0

# DEP-PY-MLFLOW-001: mlflow < 2.9.2
mlflow==2.8.0
